requirements.txt
# 대회 제공 패키지 (버전 고정)
albumentations==1.4.0
numpy==1.26.4
pandas==2.1.4
Pillow==9.4.0
timm==0.9.12
matplotlib-inline==0.1.6

# PyTorch (CUDA 필수)
# 설치: pip install torch==2.1.0 torchvision==0.16.0 --index-url https://download.pytorch.org/whl/cu118
torch==2.1.0
torchvision==0.16.0

# 실험 관리/설정
hydra-core==1.3.2
hydra-colorlog==1.2.0
pydantic==2.7.4
wandb==0.17.9
boto3==1.34.162
scikit-learn==1.4.2
pyyaml==6.0.2
python-dotenv==1.0.1

opencv-python-headless==4.10.0.84
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
ipykernel>=6.27.0
ipywidgets>=8.1.0
